+++ Credixware/Internal/Headers/Offsets.h
@@ -8,13 +8,20 @@ namespace Offsets {
 	DWORD m_fFlags = 0x104;
 	DWORD m_iHealth = 0x100;
 	DWORD m_iTeamNum = 0xF4;
+	DWORD m_iGlowIndex = 0xA3F8;
 	DWORD m_iItemDefinitionIndex = 0x2FAA;
 	DWORD m_iItemIDHigh = 0x2FC0;
 	DWORD m_nFallbackPaintKit = 0x31B8;
 	DWORD m_nFallbackSeed = 0x31BC;
 	DWORD m_nFallbackStatTrak = 0x31C4;
+	DWORD m_flFallbackWear = 0x31C0;
 	DWORD m_hMyWeapons = 0x2DF8;
+	DWORD m_hMyWearables = 0x2EF4;
 	DWORD m_hActiveWeapon = 0x2EF8;
+	DWORD dwGlowObjectManager = 0x520DAE0;
+	DWORD dwPlayerResource = 0x30FD71C;
+	DWORD m_iCompetitiveRanking = 0x1A84;
+	DWORD m_iCompetitiveWins = 0x1B88;
 }
 
 #endif