+++ Credixware/Internal/Hooks/OverrideView.h
@@ -11,8 +11,8 @@ unsigned int* oOverrideViewFunc;
 
 void __fastcall hkOverrideView(void* ecx, void* edx, CViewSetup* setup) {
 	Thirdperson::Update(setup);
-	// Uncomment for FOV Changer (NOTE: Buggy when scoped in)
-	//setup->fov = 120.0f;
+	setup->fovViewmodel = 100.0f;
+	//setup->fov = 110.0f;
 	oOverrideView(ecx, setup);
 }