+++ Credixware/Internal/SDK/SDK.h
@@ -8,6 +8,7 @@
 #include "C_BasePlayer.h"
 #include "C_BaseViewModel.h"
 #include "CGlobalVars.h"
+#include "CGlowObjectManager.h"
 #include "CInput.h"
 #include "ClientClass.h"
 #include "Color.h"